requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
environs==10.0.0
idna==3.6
marshmallow==3.20.1
packaging==23.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
